requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
wbgapi>=1.0.12
fredapi>=0.5.1
python-dotenv>=1.0.0
pyyaml>=6.0.1
sqlalchemy>=2.0.0
streamlit>=1.31.0
plotly>=5.18.0
joblib>=1.3.0
scipy>=1.11.0
